@@ -1,68 +1,66 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 5.2.6.3'
+ruby '2.7.7'
+
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails', branch: 'main'
+gem 'rails', '~> 6.1.0'
  # Use pg as the database for Active Record
-gem 'pg', '~> 0.18'
+gem 'pg'
 # Use Puma as the app server
-gem 'puma', '~> 5.3'
+gem 'puma', '~> 5.0'
 # Use SCSS for stylesheets
-gem 'sassc-rails', '~> 2.1.2'
-gem 'sprockets', '< 4'
-# Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
-# See https://github.com/rails/execjs#readme for more supported runtimes
-# gem 'mini_racer', platforms: :ruby
+gem 'sass-rails', '>= 6'
 # Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 5.0'
+gem 'coffee-rails'
+# Transpile app-like JavaScript. Read more: https://github.com/rails/webpacker
+gem 'webpacker', '~> 5.0'
 # Turbolinks makes navigating your web application faster. Read more: https://github.com/turbolinks/turbolinks
 gem 'turbolinks', '~> 5'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.11'
+# gem 'jbuilder', '~> 2.7'
 # Use Redis adapter to run Action Cable in production
 # gem 'redis', '~> 4.0'
-# Use ActiveModel has_secure_password
+# Use Active Model has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
-# Use ActiveStorage variant
-# gem 'mini_magick', '~> 4.8'
-# Use Capistrano for deployment
-# gem 'capistrano-rails', group: :development
+
+# Use Active Storage variant
+# gem 'image_processing', '~> 1.2'
 
 # Reduces boot times through caching; required in config/boot.rb
-gem 'bootsnap', '>= 1.1.0', require: false
+gem 'bootsnap', '>= 1.4.4', require: false
 
-gem 'active_model_serializers', '~> 0.10.12'
-gem 'activeadmin', '~> 1.4.3'
+gem 'active_model_serializers'
+gem 'activeadmin'
 gem 'activeadmin_addons'
-gem 'activeadmin_simplemde' # this depends on AA 1
+gem 'activeadmin_simplemde'
 gem 'activeadmin_sortable_table'
 gem 'acts_as_list'
 gem 'appsignal'
-gem 'aws-sdk-s3', '~> 1'
-gem 'best_in_place'
-gem 'ckeditor', '4.3.0'
-gem 'devise', '~> 4.7'
+gem 'aws-sdk-s3'
+gem 'best_in_place', git: 'https://github.com/mmotherwell/best_in_place' # for Rails 6.1
+gem 'ckeditor', '< 5'
+gem 'devise'
 gem 'devise_token_auth'
-gem 'dotenv-rails', '~> 2.7'
-gem 'enumerate_it', '~> 3.2.0'
-gem 'faraday'
+gem 'dotenv-rails'
+gem 'enumerate_it'
+gem 'faraday', '< 2'
 gem 'i18n_data'
 gem 'kaminari'
 gem 'mailchimp-api', require: 'mailchimp'
-# gem 'newrelic_rpm'
+# gem 'net-http' # https://github.com/ruby/net-imap/issues/16#issuecomment-803086765
 gem 'oj'
-gem 'paperclip', '~> 6.1.0'
+gem 'paperclip'
 gem 'pg_csv'
 gem 'pg_search'
-gem 'rack', '~> 2.2.4' # issue with sidekiq web in version 2.1.1, still doesn't work on sandbox in version 2.1.2
-gem 'rack-cors', '~> 1.1'
+gem 'rack'
+gem 'rack-cors'
 gem 'scenic'
-gem 'sidekiq', '~> 6.2.1'
+gem 'sidekiq', '< 7'
 gem 'sidekiq-unique-jobs'
 gem 'sitemap_generator'
 gem 'staccato'
-gem 'twitter', '~> 7.0'
+gem 'twitter'
 gem 'zip_tricks'
 gem 'wet-health_endpoint'
 gem 'whenever', require: false
