@@ -5,10 +5,10 @@ Rails.application.config.assets.version = '1.0'
 
 # Add additional assets to the asset load path.
 # Rails.application.config.assets.paths << Emoji.images_path
-# Add node_modules folder to the asset load path.
+# Add Yarn node_modules folder to the asset load path.
 Rails.application.config.assets.paths << Rails.root.join('node_modules')
 
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in the app/assets
 # folder are already added.
-Rails.application.config.assets.precompile += %w( active_admin.js active_admin.css )
+Rails.application.config.assets.precompile += %w( active_admin.js active_admin.css ckeditor/config.js)
